Fixed dropping items by player (needs testing, I don't think it works 100%) Fixed not doing damage if targetting more than 7 mobs. Fixed stacking items not stacking. Currently checked purely via inventory type.

+++ src/client/character/skills/AttackInfo.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class AttackInfo {
     public boolean fieldKey;
     public byte hits;
-    public byte mobCount;
+    public int mobCount;
     public int skillId;
     public byte slv;
     public int keyDown;
